fix: resolve exec startup host before clamp warning
